Frostfang Fury encounter: real NPC locomotion, combat state, goals gate, overhead badge
Everything below is verified against the real 2014 SOE captures (byte-exact
AddNpc/AttackProcessed/AddNotifications parses) and IDA decompilation of the
2014-03-13 client. Live-tested: full encounter loop with running wolves,
attack animations, draining health bars, damage numbers, and clean exit.

NPC movement/animation (the sliding fix):
- Walk/Run/StandAnimId default -1 ("use the model's own clips") - the live
  server sends -1 on 370/370 AddNpc packets; overriding killed locomotion
- MovementType 2 (PHYSICS) like every walking NPC in the live capture;
  locomotion anims now play client-side (no more SetAnimation hacks)
- op125 rotation = facing unit-vector; PlayerUpdatePacketExpectedSpeed (sub23)

Combat feedback:
- CombatPacketAttackProcessed field mapping locked to live dump: Guid1=Guid2=
  target, Guid3=attacker, Int5=current HP after hit (the moving health bar)
- Wolf bites use AttackProcessed (incoming numbers + recoil on the player)
- World-combat state on hit (op41 sub132/133) with 6s out-of-combat decay -
  opens the client's floating-damage-number gate; job-lock releases after
- AddNpc NameColor float->int (client reads int ARGB; float mangled the bits)
- EnemyStatus flag (AddNpc bool38) - set on every live attackable hostile

Minigame/goals system:
- EncounterDetailsResponsePacket Launch flag (trailing byte): true = client
  creates its MiniGameState (sub_9BB2D0) - the master gate that silently
  drops ALL op45 objective packets while absent; sent at arena entry
  (the real server does exactly this - resends details flag=1 on entry)
- PlayerUpdatePacketAddNotifications (op35/sub10): overhead notification
  badge (crossed-swords image 24 above the Frostfang Growler)
- PacketEncounterDataCommon (op62) combat ruleset + defaults for exit
- MiniGameStateRemovePacket (op39/sub19) clean minigame teardown
- ObjectivePackets (op45 sub5/sub1/sub2) goals panel add/activate/update
- CombatPacketEnableBossDisplay (op32/sub9)

Arena: wave pacing with 30k HP wolves / 120k Alpha (multi-hit fights like the
reference video), mixed wolf skins, surround-slot chase AI, gloam sky, boss
phase with red nameplate, victory + return-home exit sequence.

+++ src/Sanctuary.Packet/BasePlayerUpdatePacket/PlayerUpdatePacketAddNpc.cs
@@ -103,7 +103,9 @@ public class PlayerUpdatePacketAddNpc : BasePlayerUpdatePacket, ISerializablePac
     [JsonConverter(typeof(Vector4JsonConverter))]
     public Vector4 Tilt;
 
-    public float NameColor;
+    // ARGB int (client reads m_nNameColor as int32; live hostiles send 0xFFFF0000 = opaque red).
+    // Was declared float here — an int color assigned through a float conversion mangles the bits.
+    public int NameColor;
 
     public int AreaDefinitionId;
